+++ src/__tests__/parsers/parser_nodes.test.js
@@ -39,6 +39,17 @@ describe('ParsedNode', () => {
     filtered = c1_1.filter((n) => n.type === 'it', true);
     expect(filtered.length).toEqual(1);
   });
+  it('can attach lastProperty for NamedBlock', () => {
+    const itBlock = new ItBlock('abc.js', 'a test');
+    itBlock.lastProperty = 'each';
+    expect(itBlock.lastProperty).toEqual('each');
+    expect(itBlock.name).toEqual('a test');
+
+    const descBlock = new DescribeBlock('abc.js', 'a describe');
+    descBlock.lastProperty = 'only';
+    expect(descBlock.lastProperty).toEqual('only');
+    expect(descBlock.name).toEqual('a describe');
+  });
   it('throws an error when adding an unknown child', async () => {
     const root = new ParsedNode('describe', 'a/b/c');
     expect(() => {