+++ src/parsers/parser_nodes.js
@@ -104,6 +104,15 @@ export class NamedBlock extends ParsedNode {
 
   nameRange: ParsedRange;
 
+  lastProperty: ?string;
+
+  /**
+   * type of the name, it's the babel Node["type"], such as "Literal", "TemplateLiteral" etc
+   *
+   * TODO babel parser currently returns "Literal" for the it/describe name argument, which is not part of its "type" definition, therefore declare a string type for now until it is fixed in babel.
+   * */
+  nameType: ?string;
+
   constructor(type: ParsedNodeType, file: string, name?: string) {
     super(type, file);
     if (name) {